requirements/development.txt
-r base.txt
django-extensions==2.2.5
django-debug-toolbar==2.0